requirements.txt
asgiref==3.7.2
boto3==1.28.39
botocore==1.31.39
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.11.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
whitenoise==6.5.0
